pyproject.toml
[project]
name = "ktb-ai-week-langchain"
version = "0.1.0"
description = "AI-based culture-fit matching service for developer job seekers"
requires-python = ">=3.10"
license = "MIT"

dependencies = [
    # LangChain
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.0",

    # Database
    "pymongo>=4.0.0",

    # Web Scraping
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "httpx>=0.27.0",

    # Utilities
    "python-dotenv>=1.0.0",

    # AWS S3
    "boto3>=1.34.0",

    # Google Gemini (Files API)
    "google-genai>=1.0.0",

    # FastAPI Server
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
]
browser = [
    "playwright>=1.40.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["langchain_pipeline*", "api*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
